+++ model/src/progress.ts
@@ -1,4 +1,4 @@
 export const ProgressPrefix = '[==PROGRESS==]';
 
-export const ProgressPattern =
-  /(?<stage>[^:]*)\:(?: *(?<progress>[0-9.]+)\%)?(?: *ETA\: *(?<eta>.+))?/;
+export const ProgressPattern
+  = /(?<stage>[^:]*):(?: *(?<progress>[0-9.]+)%)?(?: *ETA: *(?<eta>.+))?/;