+++ ui/src/MainPage.vue
@@ -2,7 +2,8 @@
 import { AgGridVue } from 'ag-grid-vue3';
 
 import { ClientSideRowModelModule } from 'ag-grid-enterprise';
-import { ColDef, GridApi, GridOptions, GridReadyEvent, ModuleRegistry } from 'ag-grid-enterprise';
+import type { ColDef, GridApi, GridOptions, GridReadyEvent } from 'ag-grid-enterprise';
+import { ModuleRegistry } from 'ag-grid-enterprise';
 import type { PlId, Qc } from '@platforma-open/milaboratories.mixcr-clonotyping.model';
 import {
   AgGridTheme,
@@ -15,24 +16,25 @@ import {
   PlAgTextAndButtonCell,
   PlAgCellStatusTag,
   makeRowNumberColDef,
-  autoSizeRowNumberColumn
+  autoSizeRowNumberColumn,
 } from '@platforma-sdk/ui-vue';
 import { refDebounced, whenever } from '@vueuse/core';
 import { reactive, shallowRef, watch } from 'vue';
 import { useApp } from './app';
-import ProgressCell from './ProgressCell.vue';
-import { MiXCRResult, MiXCRResultsFull } from './results';
+import type { MiXCRResult } from './results';
+import { MiXCRResultsFull } from './results';
 import SampleReportPanel from './SampleReportPanel.vue';
 import SettingsPanel from './SettingsPanel.vue';
 import { getAlignmentChartSettings } from './charts/alignmentChartSettings';
 import { getChainsChartSettings } from './charts/chainsChartSettings';
-import { PlAgChartStackedBarCell } from '@platforma-sdk/ui-vue';
+import { PlAgChartStackedBarCell, createAgGridColDef } from '@platforma-sdk/ui-vue';
+import { parseProgressString } from './parseProgress';
 
 const app = useApp();
 
 // @TODO
 const result = refDebounced(MiXCRResultsFull, 100, {
-  maxWait: 200
+  maxWait: 200,
 });
 
 const data = reactive<{
@@ -42,7 +44,7 @@ const data = reactive<{
 }>({
   settingsOpen: app.model.outputs.started === false,
   sampleReportOpen: false,
-  selectedSample: undefined
+  selectedSample: undefined,
 });
 
 watch(
@@ -50,23 +52,23 @@ watch(
   (newVal, oldVal) => {
     if (oldVal === false && newVal === true) data.settingsOpen = false;
     if (oldVal === true && newVal === false) data.settingsOpen = true;
-  }
+  },
 );
 
 whenever(
   () => data.settingsOpen,
-  () => (data.sampleReportOpen = false)
+  () => (data.sampleReportOpen = false),
 );
 whenever(
   () => data.sampleReportOpen,
-  () => (data.settingsOpen = false)
+  () => (data.settingsOpen = false),
 );
 
 ModuleRegistry.registerModules([ClientSideRowModelModule]);
 
 const qcPriority = { OK: 0, WARN: 1, ALERT: 2 };
 
-const gridApi = shallowRef<GridApi<any>>();
+const gridApi = shallowRef<GridApi>();
 const onGridReady = (params: GridReadyEvent) => {
   gridApi.value = params.api;
   autoSizeRowNumberColumn(params.api);
@@ -75,10 +77,10 @@ const onGridReady = (params: GridReadyEvent) => {
 const defaultColumnDef: ColDef = {
   suppressHeaderMenuButton: true,
   lockPinned: true,
-  sortable: false
+  sortable: false,
 };
 
-const columnDefs: ColDef[] = [
+const columnDefs: ColDef<MiXCRResult>[] = [
   makeRowNumberColDef(),
   {
     colId: 'label',
@@ -89,52 +91,61 @@ const columnDefs: ColDef[] = [
     sortable: true,
     cellRenderer: PlAgTextAndButtonCell,
     cellRendererParams: {
-      invokeRowsOnDoubleClick: true
-    }
+      invokeRowsOnDoubleClick: true,
+    },
   },
-  {
+  createAgGridColDef<MiXCRResult, string>({
     colId: 'progress',
     field: 'progress',
-    cellRenderer: 'ProgressCell',
     headerName: 'Progress',
-    cellStyle: {
-      '--ag-cell-horizontal-padding': '0px',
-      '--ag-cell-vertical-padding': '0px'
-    }
-  },
-  {
+    progress(cellData) {
+      const parsed = parseProgressString(cellData.value);
+
+      if (parsed.stage === 'Queued') {
+        return {
+          status: 'not_started',
+          text: parsed.stage,
+        };
+      }
+
+      return {
+        status: parsed.stage === 'Done' ? 'done' : 'running',
+        percent: parsed.percentage,
+        text: parsed.stage,
+        suffix: parsed.etaLabel ?? '',
+      };
+    },
+  }),
+  createAgGridColDef({
     colId: 'qc',
     field: 'qc',
     width: 96,
     cellRendererSelector: (cellData) => {
-      const type = (cellData.data.qc as MiXCRResult['qc'])?.reduce(
+      const type = (cellData.data?.qc as MiXCRResult['qc'])?.reduce(
         (result: Qc[number]['status'], item) =>
           qcPriority[item.status] > qcPriority[result] ? item.status : result,
-        'OK'
+        'OK',
       );
       return {
         component: PlAgCellStatusTag,
-        params: { type }
+        params: { type },
       };
     },
     headerName: 'Quality',
-    cellStyle: {
-      '--ag-cell-horizontal-padding': '0px',
-      '--ag-cell-vertical-padding': '0px'
-    }
-  },
+    noGutters: true, // this means "no padding" i. e. --ag-cell-horizontal-padding: 0px & --ag-cell-vertical-padding: 0px
+  }),
   {
     colId: 'alignmentStats',
     headerName: 'Alignments',
     flex: 1,
     cellStyle: {
-      '--ag-cell-horizontal-padding': '12px'
+      '--ag-cell-horizontal-padding': '12px',
     },
     cellRendererSelector: (cellData) => {
-      const value = getAlignmentChartSettings(cellData.data.alignReport);
+      const value = getAlignmentChartSettings(cellData.data?.alignReport);
       return {
         component: PlAgChartStackedBarCell,
-        params: { value }
+        params: { value },
       };
     },
   },
@@ -143,23 +154,19 @@ const columnDefs: ColDef[] = [
     headerName: 'Chains',
     flex: 1,
     cellStyle: {
-      '--ag-cell-horizontal-padding': '12px'
+      '--ag-cell-horizontal-padding': '12px',
       // '--ag-cell-horizontal-border': 'solid rgb(150, 150, 200);',
       // 'border-width': '0'
     },
     cellRendererSelector: (cellData) => {
-      const value = getChainsChartSettings(cellData.data.alignReport);
+      const value = getChainsChartSettings(cellData.data?.alignReport);
       return {
         component: PlAgChartStackedBarCell,
-        params: { value }
+        params: { value },
       };
     },
-  }
+  },
 ];
-
-// watch(result, rd => {
-//     console.dir(rd, { depth: 5 })
-// }, { immediate: true })
 
 const gridOptions: GridOptions<MiXCRResult> = {
   getRowId: (row) => row.data.sampleId,
@@ -168,9 +175,8 @@ const gridOptions: GridOptions<MiXCRResult> = {
     data.sampleReportOpen = data.selectedSample !== undefined;
   },
   components: {
-    ProgressCell,
-    PlAgTextAndButtonCell
-  }
+    PlAgTextAndButtonCell,
+  },
 };
 </script>
 
@@ -217,7 +223,7 @@ const gridOptions: GridOptions<MiXCRResult> = {
       Results for
       {{
         (data.selectedSample ? app.model.outputs.sampleLabels?.[data.selectedSample] : undefined) ??
-        '...'
+          '...'
       }}
     </template>
     <SampleReportPanel v-model="data.selectedSample" />