+++ model/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.model
 
+## 1.7.0
+
+### Minor Changes
+
+- a370729: Custom library species corrected
+
+## 1.6.0
+
+### Minor Changes
+
+- db3cc0a: Custom reference library from library builder block or file
+
+## 1.5.1
+
+### Patch Changes
+
+- 392d6f3: Add and use @platforma-sdk/eslint-config package (MILAB-1128)
+
 ## 1.5.0
 
 ### Minor Changes